+++ src/test/java/testsupport/Constants.java
@@ -1,0 +1,8 @@
+package testsupport;
+
+public final class Constants {
+    public static final Integer STACK_EXPLODING_NUMBER = 50_000;
+
+    private Constants() {
+    }
+}